+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add treatment checkbox in UI to allow adding a rotation separate from rotation as treatment. [#404](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/404)
 - Experiment plan deviations tab and make fields multivalue. [#405](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/405)
 - Change proposal statistical design to text_long. [#407](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/407)
+- Split proposal Design tab in two. [#408](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/408)
 
 ## [2.10.1](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/16) 2023-04-05